+++ pygfunction/heat_transfer.py
@@ -104,7 +104,10 @@ def finite_line_source(
             dis, H1, D1, H2, D2, reaSource, imgSource)
 
         # Evaluate integral
-        if isinstance(time, (np.floating, float)):
+        if time == np.inf:
+            h = _finite_line_source_steady_state(
+                dis, H1, D1, H2, D2, reaSource, imgSource)
+        elif isinstance(time, (np.floating, float)):
             # Lower bound of integration
             a = 1.0 / np.sqrt(4.0*alpha*time)
             h = quad(f, a, np.inf)[0]
@@ -116,17 +119,25 @@ def finite_line_source(
         # Unpack parameters
         if isinstance(borehole1, Borehole): borehole1 = [borehole1]
         if isinstance(borehole2, Borehole): borehole2 = [borehole2]
-        dis = np.array(
-            [[b1.distance(b2) for b1 in borehole1] for b2 in borehole2])
+        x1 = np.array([b.x for b in borehole1])
+        y1 = np.array([b.y for b in borehole1])
+        x2 = np.array([b.x for b in borehole2])
+        y2 = np.array([b.y for b in borehole2])
+        r_b = np.array([b.r_b for b in borehole1])
+        dis = np.maximum(np.sqrt(np.add.outer(x2, -x1)**2 + np.add.outer(y2, -y1)**2), r_b)
         D1 = np.array([b.D for b in borehole1]).reshape(1, -1)
         H1 = np.array([b.H for b in borehole1]).reshape(1, -1)
         D2 = np.array([b.D for b in borehole2]).reshape(-1, 1)
         H2 = np.array([b.H for b in borehole2]).reshape(-1, 1)
 
-        # Evaluate integral
-        h = finite_line_source_vectorized(
-            time, alpha, dis, H1, D1, H2, D2,
-            reaSource=reaSource, imgSource=imgSource)
+        if time is np.inf:
+            h = _finite_line_source_steady_state(
+                dis, H1, D1, H2, D2, reaSource, imgSource)
+        else:
+            # Evaluate integral
+            h = finite_line_source_vectorized(
+                time, alpha, dis, H1, D1, H2, D2,
+                reaSource=reaSource, imgSource=imgSource)
     return h
 
 
@@ -235,6 +246,110 @@ def finite_line_source_vectorized(
     return h
 
 
+def finite_line_source_equivalent_boreholes_vectorized(
+        time, alpha, dis, wDis, H1, D1, H2, D2, N2, reaSource=True, imgSource=True):
+    """
+    Evaluate the equivalent Finite Line Source (FLS) solution.
+
+    This function uses a numerical quadrature to evaluate the one-integral form
+    of the FLS solution, as proposed by Prieto and Cimmino
+    [#eqFLSVec-PriCim2021]_. The equivalent FLS solution is given by:
+
+        .. math::
+            h_{1\\rightarrow2}(t) &= \\frac{1}{2 H_2 N_{b,2}}
+            \\int_{\\frac{1}{\\sqrt{4\\alpha t}}}^{\\infty}
+            \\sum_{G_1} \\sum_{G_2}
+            e^{-d_{12}^2s^2}(I_{real}(s)+I_{imag}(s))ds
+
+
+            I_{real}(s) &= erfint((D_2-D_1+H_2)s) - erfint((D_2-D_1)s)
+
+            &+ erfint((D_2-D_1-H_1)s) - erfint((D_2-D_1+H_2-H_1)s)
+
+            I_{imag}(s) &= erfint((D_2+D_1+H_2)s) - erfint((D_2+D_1)s)
+
+            &+ erfint((D_2+D_1+H_1)s) - erfint((D_2+D_1+H_2+H_1)s)
+
+
+            erfint(X) &= \\int_{0}^{X} erf(x) dx
+
+                      &= Xerf(X) - \\frac{1}{\\sqrt{\\pi}}(1-e^{-X^2})
+
+        .. Note::
+            The reciprocal thermal response factor
+            :math:`h_{2\\rightarrow1}(t)` can be conveniently calculated by:
+
+                .. math::
+                    h_{2\\rightarrow1}(t) = \\frac{H_2 N_{b,2}}{H_1 N_{b,1}}
+                    h_{1\\rightarrow2}(t)
+
+    Parameters
+    ----------
+    time : float or array, shape (K)
+        Value of time (in seconds) for which the FLS solution is evaluated.
+    alpha : float
+        Soil thermal diffusivity (in m2/s).
+    dis : array
+        Unique radial distances to evaluate the FLS solution.
+    wDis : array
+        Number of instances of each unique radial distances.
+    H1 : float or array
+        Lengths of the emitting heat sources.
+    D1 : float or array
+        Buried depths of the emitting heat sources.
+    H2 : float or array
+        Lengths of the receiving heat sources.
+    D2 : float or array
+        Buried depths of the receiving heat sources.
+    reaSource : bool
+        True if the real part of the FLS solution is to be included.
+        Default is True.
+    imgSource : bool
+        True if the image part of the FLS solution is to be included.
+        Default is True.
+
+    Returns
+    -------
+    h : float
+        Value of the FLS solution. The average (over the length) temperature
+        drop on the wall of borehole2 due to heat extracted from borehole1 is:
+
+        .. math:: \\Delta T_{b,2} = T_g - \\frac{Q_1}{2\\pi k_s H_2} h
+
+    Notes
+    -----
+    This is a vectorized version of the :func:`finite_line_source` function
+    using scipy.integrate.quad_vec to speed up calculations. All arrays
+    (dis, H1, D1, H2, D2) must follow numpy array broadcasting rules. If time
+    is an array, the integrals for different time values are stacked on the
+    last axis.
+    
+
+    References
+    ----------
+    .. [#eqFLSVec-PriCim2021] Prieto, C., & Cimmino, M.
+       (2021). Thermal interactions in large irregular fields of geothermal
+       boreholes: the method of equivalent borehole. Journal of Building
+       Performance Simulation, 14 (4), 446-460.
+
+    """
+    # Integrand of the finite line source solution
+    f = _finite_line_source_equivalent_boreholes_integrand(
+        dis, wDis, H1, D1, H2, D2, N2, reaSource, imgSource)
+
+    # Evaluate integral
+    if isinstance(time, (np.floating, float)):
+        # Lower bound of integration
+        a = 1.0 / np.sqrt(4.0*alpha*time)
+        h = quad_vec(f, a, np.inf)[0]
+    else:
+        h = np.stack(
+            [quad_vec(f, 1.0 / np.sqrt(4.0*alpha*t), np.inf)[0]
+             for t in time],
+            axis=-1)
+    return h
+
+
 def _erfint(x):
     """
     Integral of the error function.
@@ -320,3 +435,147 @@ def _finite_line_source_integrand(dis, H1, D1, H2, D2, reaSource, imgSource):
         # No heat source
         f = lambda s: 0.
     return f
+
+
+def _finite_line_source_equivalent_boreholes_integrand(dis, wDis, H1, D1, H2, D2, N2, reaSource, imgSource):
+    """
+    Integrand of the finite line source solution.
+
+    Parameters
+    ----------
+    dis : array
+        Unique radial distances to evaluate the FLS solution.
+    wDis : array
+        Number of instances of each unique radial distances.
+    H1 : float or array
+        Lengths of the emitting heat sources.
+    D1 : float or array
+        Buried depths of the emitting heat sources.
+    H2 : float or array
+        Lengths of the receiving heat sources.
+    D2 : float or array
+        Buried depths of the receiving heat sources.
+    N2 :
+    reaSource : bool
+        True if the real part of the FLS solution is to be included.
+    imgSource : bool
+        True if the image part of the FLS solution is to be included.
+
+    Returns
+    -------
+    f : callable
+        Integrand of the finite line source solution. Can be vector-valued.
+
+    Notes
+    -----
+    All arrays (dis, H1, D1, H2, D2) must follow numpy array broadcasting
+    rules.
+
+    """
+    if reaSource and imgSource:
+        # Full (real + image) FLS solution
+        p = np.array([1, -1, 1, -1, 1, -1, 1, -1])
+        q = np.stack([D2 - D1 + H2,
+                      D2 - D1,
+                      D2 - D1 - H1,
+                      D2 - D1 + H2 - H1,
+                      D2 + D1 + H2,
+                      D2 + D1,
+                      D2 + D1 + H1,
+                      D2 + D1 + H2 + H1],
+                     axis=-1)
+        f = lambda s: 0.5 / (N2*H2*s**2) * (np.exp(-dis**2*s**2) @ wDis).T * np.inner(p, _erfint(q*s))
+    elif reaSource:
+        # Real FLS solution
+        p = np.array([1, -1, 1, -1])
+        q = np.stack([D2 - D1 + H2,
+                      D2 - D1,
+                      D2 - D1 - H1,
+                      D2 - D1 + H2 - H1],
+                     axis=-1)
+        f = lambda s: 0.5 / (N2*H2*s**2) * (np.exp(-dis**2*s**2) @ wDis).T * np.inner(p, _erfint(q*s))
+    elif imgSource:
+        # Image FLS solution
+        p = np.array([1, -1, 1, -1])
+        q = np.stack([D2 + D1 + H2,
+                      D2 + D1,
+                      D2 + D1 + H1,
+                      D2 + D1 + H2 + H1],
+                     axis=-1)
+        f = lambda s: 0.5 / (N2*H2*s**2) * (np.exp(-dis**2*s**2) @ wDis).T * np.inner(p, _erfint(q*s))
+    else:
+        # No heat source
+        f = lambda s: 0.
+    return f
+
+
+def _finite_line_source_steady_state(dis, H1, D1, H2, D2, reaSource, imgSource):
+    """
+    Integrand of the finite line source solution.
+
+    Parameters
+    ----------
+    dis : float or array
+        Radial distances to evaluate the FLS solution.
+    H1 : float or array
+        Lengths of the emitting heat sources.
+    D1 : float or array
+        Buried depths of the emitting heat sources.
+    H2 : float or array
+        Lengths of the receiving heat sources.
+    D2 : float or array
+        Buried depths of the receiving heat sources.
+    reaSource : bool
+        True if the real part of the FLS solution is to be included.
+    imgSource : bool
+        True if the image part of the FLS solution is to be included.
+
+    Returns
+    -------
+    h : Steady-state finite line source solution.
+
+    Notes
+    -----
+    All arrays (dis, H1, D1, H2, D2) must follow numpy array broadcasting
+    rules.
+
+    """
+    # Steady-state solution
+    if reaSource and imgSource:
+        # Full (real + image) FLS solution
+        p = np.array([1, -1, 1, -1, 1, -1, 1, -1])
+        q = np.stack([D2 - D1 + H2,
+                      D2 - D1,
+                      D2 - D1 - H1,
+                      D2 - D1 + H2 - H1,
+                      D2 + D1 + H2,
+                      D2 + D1,
+                      D2 + D1 + H1,
+                      D2 + D1 + H2 + H1],
+                      axis=-1)
+        dis = np.expand_dims(dis, axis=-1)
+        h = 0.5 / H2 * np.inner(p, q * np.log((q + np.sqrt(q**2 + dis**2)) / dis) - np.sqrt(q**2 + dis**2))
+    elif reaSource:
+        # Real FLS solution
+        p = np.array([1, -1, 1, -1])
+        q = np.stack([D2 - D1 + H2,
+                      D2 - D1,
+                      D2 - D1 - H1,
+                      D2 - D1 + H2 - H1,],
+                      axis=-1)
+        dis = np.expand_dims(dis, axis=-1)
+        h = 0.5 / H2 * np.inner(p, q * np.log((q + np.sqrt(q**2 + dis**2)) / dis) - np.sqrt(q**2 + dis**2))
+    elif imgSource:
+        # Image FLS solution
+        p = np.array([1, -1, 1, -1])
+        q = np.stack([D2 + D1 + H2,
+                      D2 + D1,
+                      D2 + D1 + H1,
+                      D2 + D1 + H2 + H1],
+                      axis=-1)
+        dis = np.expand_dims(dis, axis=-1)
+        h = 0.5 / H2 * np.inner(p, q * np.log((q + np.sqrt(q**2 + dis**2)) / dis) - np.sqrt(q**2 + dis**2))
+    else:
+        # No heat source
+        h = 0.
+    return h